[Refactor]getRoutineElementsByTypeを平行棒の制限ルールに適用
