+++ frontend/src/utilities/ElementUtil.ts
@@ -1,4 +1,4 @@
-import { difficulties, getEventKey, getGroupKey } from "./Type";
+import { difficulties, ElementGroup, ElementType, Events, getEventKey, getGroupKey } from "./Type";
 
 export interface Element {
   id: number;
@@ -123,4 +123,24 @@ export const isElementTypeIncluded = (elementTypes: string | null, elementTypeTo
     return false;
   }
   return elementTypes.split("-").includes(elementTypeToCheck.toString());
+};
+
+// 指定のtypeに該当するElementを取得
+export const getElementsByType = (
+  selectEvent: Events,
+  selectGroup: ElementGroup,
+  targetElementType: ElementType,
+  categorizedElements: CategorizedElements
+): Element[] => {
+  const groupElements = getGroupElements(categorizedElements, selectEvent, selectGroup);
+
+  const felgeElements = Object.values(groupElements).flatMap((rowElements) =>
+    Object.values(rowElements).filter(
+      (element) =>
+        "element_type" in element && isElementTypeIncluded(element.element_type, targetElementType)
+    )
+  ) as Element[];
+
+  // 技のコードを昇順に並べる
+  return felgeElements.sort((a, b) => a.id - b.id);
 };
