+++ frontend/src/utilities/RoutinePBUtil.ts
@@ -1,15 +1,9 @@
-import {
-  CategorizedElements,
-  Element,
-  getGroupElements,
-  isElementTypeIncluded,
-} from "./ElementUtil";
+import { Element, isElementTypeIncluded } from "./ElementUtil";
 import { RoutineElement } from "./RoutineUtil";
 import {
   ElementGroup,
   ElementStatus,
   ElementType,
-  Events,
   getElementStatusFromElementType,
   getElementTypeName,
 } from "./Type";
@@ -72,65 +66,4 @@ export const getPBSaltoLimitCodes = (routine: RoutineElement[]) => {
     });
 
   return codes;
-};
-
-// RoutineRules用 | 車輪系制限コード取得
-export const getPBGiantSwingLimitCodes = (routine: RoutineElement[]) => {
-  let codes: { id: number; code: string }[] = [];
-  routine
-    .filter((element) => element.is_qualified === true)
-    .forEach((element) => {
-      if (isElementTypeIncluded(element.element_type, ElementType.平行棒_車輪系)) {
-        codes.push({ id: element.id!, code: element.code! });
-      }
-    });
-
-  return codes;
-};
-
-// RoutineRules用 | 棒下宙返り系制限コード取得
-export const getPBFelgeLimitCodes = (routine: RoutineElement[]) => {
-  let codes: { id: number; code: string }[] = [];
-  routine
-    .filter((element) => element.is_qualified === true)
-    .forEach((element) => {
-      if (isElementTypeIncluded(element.element_type, ElementType.平行棒_棒下宙返り系)) {
-        codes.push({ id: element.id!, code: element.code! });
-      }
-    });
-
-  return codes;
-};
-
-// RoutineRules用 | 棒下宙返り系の技を取得
-export const getElementsByType = (
-  selectEvent: Events,
-  selectGroup: ElementGroup,
-  targetElementType: ElementType,
-  categorizedElements: CategorizedElements
-): Element[] => {
-  const groupElements = getGroupElements(categorizedElements, selectEvent, selectGroup);
-
-  const felgeElements = Object.values(groupElements).flatMap((rowElements) =>
-    Object.values(rowElements).filter(
-      (element) =>
-        "element_type" in element && isElementTypeIncluded(element.element_type, targetElementType)
-    )
-  ) as Element[];
-
-  // 技のコードを昇順に並べる
-  return felgeElements.sort((a, b) => a.id - b.id);
-};
-
-// RoutineRules用 | アーム倒立系制限技を取得
-export const getPBFrontUpriseLimitCodes = (routine: RoutineElement[]) => {
-  let codes: { id: number; code: string }[] = [];
-  routine
-    .filter((element) => element.is_qualified === true)
-    .forEach((element) => {
-      if (isElementTypeIncluded(element.element_type, ElementType.平行棒_アーム倒立系)) {
-        codes.push({ id: element.id!, code: element.code! });
-      }
-    });
-  return codes.sort((a, b) => a.id - b.id);
 };
